# apps/messaging-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO Messaging API
  description: API for OGCIO Messaging Service
  version: 0.1.0
components:
  schemas:
    def-0:
      type: object
      properties:
        statusCode:
          type: number
        code:
          type: string
        error:
          type: string
        message:
          type: string
      title: HttpError
paths:
...
  /api/v1/templates/:
    get:
      tags: &a4
        - Templates
      parameters:
        - schema:
            type: string
          in: query
          name: lang
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        templateMetaId:
                          format: uuid
                          type: string
                        lang:
                          type: string
                        templateName:
                          type: string
                      required:
                        - templateMetaId
                        - lang
                        - templateName
                required:
                  - data
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    post:
      tags: *a4
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contents:
                  type: array
                  items:
                    type: object
                    properties:
                      templateName:
                        type: string
                      lang:
                        type: string
                      subject:
                        type: string
                      excerpt:
                        type: string
                      plainText:
                        type: string
                      richText:
                        type: string
                    required:
                      - templateName
                      - lang
                      - subject
                      - excerpt
                      - plainText
                      - richText
                variables:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
                      languages:
                        type: array
                        items:
                          type: string
                    required:
                      - name
                      - type
                      - languages
              required:
                - contents
                - variables
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        format: uuid
                        type: string
                    required:
                      - id
                required:
                  - data
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/templates/{templateId}:
    get:
      tags: *a4
      parameters:
        - schema:
            type: string
          in: path
          name: templateId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      contents:
                        type: array
                        items:
                          type: object
                          properties:
                            templateName:
                              type: string
                            subject:
                              type: string
                            excerpt:
                              type: string
                            plainText:
                              type: string
                            richText:
                              type: string
                            lang:
                              type: string
                          required:
                            - templateName
                            - subject
                            - excerpt
                            - plainText
                            - richText
                            - lang
                      fields:
                        type: array
                        items:
                          type: object
                          properties:
                            fieldName:
                              type: string
                            fieldType:
                              type: string
                          required:
                            - fieldName
                            - fieldType
                    required:
                      - contents
                      - fields
                required:
                  - data
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    put:
      tags: *a4
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contents:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        format: uuid
                        type: string
                      templateName:
                        type: string
                      lang:
                        type: string
                      subject:
                        type: string
                      excerpt:
                        type: string
                      plainText:
                        type: string
                      richText:
                        type: string
                    required:
                      - id
                      - templateName
                      - lang
                      - subject
                      - excerpt
                      - plainText
                      - richText
                variables:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
                    required:
                      - name
                      - type
              required:
                - contents
                - variables
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: templateId
          required: true
      responses:
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    delete:
      tags: *a4
      parameters:
        - schema:
            type: string
          in: path
          name: templateId
          required: true
      responses:
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...